fix(token-manager): change millis to mins

+++ app/src/main/java/com/haykor/fridge/core/data/remote/models/AuthModels.kt
@@ -18,10 +18,10 @@ data class AuthResponse(
     val refreshToken: String,
 
     @SerialName("access_token_expires_in")
-    val accessTokenExpiresIn: Long,
+    val accessTokenExpiresIn: Int, // MINUTES
 
     @SerialName("refresh_token_expires_in")
-    val refreshTokenExpiresIn: Long
+    val refreshTokenExpiresIn: Int // MINUTES
 )
 
 @Serializable
